v0.2/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.17
httpx==0.27.2
openai==1.54.0
python-dotenv==1.0.1
alembic==1.14.0
pydantic==2.9.2
pydantic-settings==2.6.1